# contracts/auvra-v2.openapi.json
{
  "components": {
    "schemas": {
...
      "AssessmentResponse": {
        "additionalProperties": false,
        "properties": {
          "assessment_id": {
            "format": "uuid",
            "title": "Assessment Id",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "validated_at": {
            "format": "date-time",
            "title": "Validated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "assessment_id",
          "session_id",
          "version",
          "schema_version",
          "timezone",
          "validated_at"
        ],
        "title": "AssessmentResponse",
        "type": "object"
      },
      "AssessmentWriteRequest": {
        "additionalProperties": false,
        "properties": {
          "answers": {
            "$ref": "#/components/schemas/MobileQuestionnaireV1"
          },
          "schema_version": {
            "const": "mobile-questionnaire.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "timezone": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "timezone",
          "answers"
        ],
        "title": "AssessmentWriteRequest",
        "type": "object"
      },
      "ClaimOnboardingRequest": {
        "additionalProperties": false,
        "properties": {
          "consents": {
            "items": {
              "$ref": "#/components/schemas/ConsentDecision"
            },
            "maxItems": 10,
            "minItems": 2,
            "title": "Consents",
            "type": "array"
          }
        },
        "required": [
          "consents"
        ],
        "title": "ClaimOnboardingRequest",
        "type": "object"
      },
      "ClaimOnboardingResponse": {
        "additionalProperties": false,
        "properties": {
          "assessment_id": {
            "format": "uuid",
            "title": "Assessment Id",
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileResponse"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "assessment_id",
          "profile"
        ],
        "title": "ClaimOnboardingResponse",
        "type": "object"
      },
      "ConsentDecision": {
        "additionalProperties": false,
        "properties": {
          "consent_type": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Consent Type",
            "type": "string"
          },
          "document_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Document Version",
            "type": "string"
          },
          "granted": {
            "title": "Granted",
            "type": "boolean"
          }
        },
        "required": [
          "consent_type",
          "document_version",
          "granted"
        ],
        "title": "ConsentDecision",
        "type": "object"
      },
      "ConsentRequirement": {
        "additionalProperties": false,
        "description": "The server-selected consent document a claimant must decide on.",
        "properties": {
          "consent_type": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Consent Type",
            "type": "string"
          },
          "document_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Document Version",
            "type": "string"
          }
        },
        "required": [
          "consent_type",
          "document_version"
        ],
        "title": "ConsentRequirement",
        "type": "object"
      },
...
      "MobileQuestionnaireV1": {
        "additionalProperties": false,
        "description": "The documented answer keys emitted by the mobile QuestionScreen.",
        "properties": {
          "age": {
            "anyOf": [
              {
                "maximum": 120.0,
                "minimum": 13.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "birth_control": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "Hormonal Birth Control Pills",
                    "IUD (Intrauterine Device)",
                    "Copper IUD (Intrauterine Device)"
                  ],
                  "type": "string"
                },
                "maxItems": 3,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Control"
          },
          "body_concerns": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "Bloating",
                    "Hot Flashes",
                    "Nausea",
                    "Difficulty losing weight / stubborn belly fat",
                    "Recent weight gain",
                    "Menstrual headaches"
                  ],
                  "type": "string"
                },
                "maxItems": 6,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Concerns"
          },
          "cycle_length": {
            "anyOf": [
              {
                "enum": [
                  "Less than 21 days",
                  "21-25 days",
                  "26-30 days",
                  "31-35 days",
                  "35+ days",
                  "I'm not sure"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cycle Length"
          },
          "diagnosed_conditions": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                "maxItems": 14,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosed Conditions"
          },
          "family_history": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                "maxItems": 14,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family History"
          },
          "last_period_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Period Date"
          },
          "lifestyle_focus": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "eat",
                    "move",
                    "pause"
                  ],
                  "type": "string"
                },
                "maxItems": 3,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifestyle Focus"
          },
          "mental_health_concerns": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "Mood swings",
                    "Stress",
                    "Fatigue"
                  ],
                  "type": "string"
                },
                "maxItems": 3,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mental Health Concerns"
          },
          "other_concerns": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                "maxItems": 2,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Concerns"
          },
          "period_concerns": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "Irregular Periods",
                    "Painful Periods",
                    "Light periods / Spotting",
                    "Heavy periods"
                  ],
                  "type": "string"
                },
                "maxItems": 4,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Concerns"
          },
          "period_description": {
            "anyOf": [
              {
                "enum": [
                  "Regular",
                  "Irregular",
                  "Occasional Skips",
                  "I don't get periods",
                  "I'm not sure"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Description"
          },
          "skin_hair_concerns": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "Hirsutism (hair growth on chin, nipples etc)",
                    "Thinning of hair",
                    "Adult Acne"
                  ],
                  "type": "string"
                },
                "maxItems": 3,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skin Hair Concerns"
          },
          "sleep_duration": {
            "anyOf": [
              {
                "enum": [
                  "<6 hours",
                  "6-7 hours",
                  "7-8 hours",
                  "8+ hours"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sleep Duration"
          },
          "stress_level": {
            "anyOf": [
              {
                "enum": [
                  "Low",
                  "Moderate",
                  "High"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stress Level"
          },
          "top_concern": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Concern"
          },
          "workout_intensity": {
            "anyOf": [
              {
                "enum": [
                  "Low",
                  "Moderate",
                  "High",
                  "I'm yet to start"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workout Intensity"
          }
        },
        "title": "MobileQuestionnaireV1",
        "type": "object"
      },
      "OnboardingSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "proof_token": {
            "title": "Proof Token",
            "type": "string"
          },
          "required_consents": {
            "items": {
              "$ref": "#/components/schemas/ConsentRequirement"
            },
            "title": "Required Consents",
            "type": "array"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "proof_token",
          "expires_at",
          "required_consents"
        ],
        "title": "OnboardingSessionResponse",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": true,
        "description": "RFC 9457 base fields plus stable AUVRA error identifiers.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "request_id"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "ProfileResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "email",
          "email_verified",
          "display_name",
          "timezone",
          "locale",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "ProfileResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AUVRA v2 mobile API",
    "title": "AUVRA API",
    "version": "2.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v2/onboarding/sessions": {
      "post": {
        "operationId": "createOnboardingSessionV2",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Precondition Failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Precondition Required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Session",
        "tags": [
          "v2-onboarding"
        ]
      }
    },
    "/api/v2/onboarding/sessions/{session_id}/assessment": {
      "put": {
        "operationId": "putOnboardingAssessmentV2",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Onboarding-Proof",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 32,
              "title": "X-Onboarding-Proof",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Quoted strong revision ETag for this resource.",
                "schema": {
                  "pattern": "^\"[0-9]+\"$",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Precondition Failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Precondition Required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Write Assessment",
        "tags": [
          "v2-onboarding"
        ]
      }
    },
    "/api/v2/onboarding/sessions/{session_id}/claim": {
      "post": {
        "operationId": "claimOnboardingSessionV2",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Onboarding-Proof",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 32,
              "title": "X-Onboarding-Proof",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimOnboardingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimOnboardingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Precondition Failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Precondition Required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {}
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Claim Session",
        "tags": [
          "v2-onboarding"
        ]
      }
    },
...
  }
}